Refactor: 프로필 사진 필수 여부 체크 로직에 updateRequestDTO.getProfilePhoto() != null 조건 추가 (#117)

+++ src/main/java/tavebalak/OTTify/user/service/UserServiceImpl.java
@@ -220,8 +220,8 @@ public class UserServiceImpl implements UserService {
         checkNicknameDuplication(userId, updateRequestDTO);
         user.changeNickName(updateRequestDTO.getNickName());
 
-        // 프로필 사진이 존재하는 경우 프로필 사진 변경
-        if (!updateRequestDTO.getProfilePhoto().isEmpty()) {
+        // 프로필 사진이 존재하고 유효한 사진인 경우 프로필 사진 변경
+        if (updateRequestDTO.getProfilePhoto() != null && !updateRequestDTO.getProfilePhoto().isEmpty()) {
             // 이전 프로필 사진 S3에서 삭제
             awss3Service.delete(user.getProfilePhoto());
